+++ CHANGELOG.md
@@ -10,11 +10,29 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Fixed compatibility with other license generators (see below) - #18
 - **BC BREAK** Updated option names to be generator specific:
-  - `license` is now `ccLicense`.
-  - `licensePrompt` is now `ccLicensePrompt`.
-  - `work` is now `ccWork`.
+    - `license` is now `ccLicense`.
+    - `licensePrompt` is now `ccLicensePrompt`.
+    - `work` is now `ccWork`.
+- **BC BREAK** Updated structure of the generator from `generators/app` to
+  `generators`. When using this as a sub-generator, make sure to use the new
+  structure when using `.composeWith`.
+
+    **New Syntax**
+
+    ```js
+    this.composeWith(require.resolve('generator-license-cc/app'), {
+      // ....
+    }
+    ```
+    **Old Syntax**
+
+    ```js
+    this.composeWith(require.resolve('generator-license-cc/generators/app'), {
+      // ....
+    }
+    ```
 - Fixed `main` definition in `package.json` to point to
-  `generators/app/index.js`
+  `app/index.js`
 - Update `.gitattributes` file to exclude directories and files from dist.
 
 ## 1.0.1 - 2017-01-22